shifted original api tests in test_api_v1.py
